+++ vitepress/.vitepress/theme/components/PreviousMonthButton.vue
@@ -7,14 +7,14 @@ const props = defineProps<{
     unit: Unit;
   }>()
 
-const temporal = useMinuta();
-const targetPeriod = usePeriod(temporal, computed(() => props.unit));
+const minuta = useMinuta();
+const targetPeriod = usePeriod(minuta, computed(() => props.unit));
 
 
 </script>
 
 <template>
-  <button class="temporal-nav-button" @click="temporal.previous(targetPeriod)">
+  <button class="minuta-nav-button" @click="minuta.previous(targetPeriod)">
     <slot>Prev</slot>
   </button>
 </template>