+++ vitepress/.vitepress/theme/components/WeekNamesView.vue
@@ -2,11 +2,11 @@
 import { computed } from "vue";
 import { useMinuta, usePeriod } from "minuta-vue";
 
-const temporal = useMinuta();
-const weekPeriod = usePeriod(temporal, "week");
-const weekdays = computed(() => temporal.divide(weekPeriod.value, "day"));
+const minuta = useMinuta();
+const weekPeriod = usePeriod(minuta, "week");
+const weekdays = computed(() => minuta.divide(weekPeriod.value, "day"));
 
-const formatter = new Intl.DateTimeFormat(temporal.locale, { weekday: "short" });
+const formatter = new Intl.DateTimeFormat(minuta.locale, { weekday: "short" });
 
 </script>