+++ vitepress/.vitepress/theme/components/WeekView.vue
@@ -11,12 +11,12 @@ const props = defineProps<{
   week: Period | Ref<Period>;
 }>();
 
-const temporal = useMinuta();
+const minuta = useMinuta();
 const targetWeek = computed(() => unref(props.week));
-const days = computed(() => temporal.divide(targetWeek.value, "day"));
+const days = computed(() => minuta.divide(targetWeek.value, "day"));
 const hourFormatter = computed(
   () =>
-    new Intl.DateTimeFormat(temporal.locale ?? "en", {
+    new Intl.DateTimeFormat(minuta.locale ?? "en", {
       hour: "numeric",
     })
 );
@@ -24,7 +24,7 @@ const hourFormatter = computed(
 const dayColumns = computed(() =>
   days.value.map((period) => ({
     period,
-    isToday: temporal.isSame(period, temporal.now.value, "day"),
+    isToday: minuta.isSame(period, minuta.now.value, "day"),
   }))
 );